% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{darkly_sync}
\alias{darkly_sync}
\alias{get_light_theme}
\alias{get_dark_theme}
\alias{set_light}
\alias{set_dark}
\title{Synchronize RStudio theme with OS appearance}
\usage{
darkly_sync()

get_light_theme()

get_dark_theme()

set_light()

set_dark()
}
\description{
Synchronize RStudio theme with OS appearance
}
\section{Functions}{
\itemize{
\item \code{darkly_sync}: Gets the OS appearance and applies the selected RStudio editor theme

\item \code{get_light_theme}: Reads the darkly light theme

\item \code{get_dark_theme}: Reads the darkly dark theme

\item \code{set_light}: Set RStudio editor to the darkly light theme

\item \code{set_dark}: Set RStudio editor to the darkly dark theme
}}

