% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{use_darkly}
\alias{use_darkly}
\title{Install darkly helper function}
\usage{
use_darkly(light_theme = "Textmate (default)",
  dark_theme = "Solarized Dark")
}
\arguments{
\item{light_theme}{Character. Name of your preferred light theme.}

\item{dark_theme}{Character. Name of your preferred dark theme.}
}
\description{
This installs a helper function into your RProfile that will check the OS
appearance and apply your chosen light or dark theme appropriately. Preferred
themes are stored in the global .Renviron file.
}
