% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{darkly_sync}
\alias{darkly_sync}
\alias{darkly_light}
\alias{darkly_dark}
\title{Synchronize RStudio theme with OS appearance}
\usage{
darkly_sync()

darkly_light()

darkly_dark()
}
\description{
Synchronize RStudio theme with OS appearance
}
\section{Functions}{
\itemize{
\item \code{darkly_sync}: Gets the OS appearance and applys the selected RStudio editor theme

\item \code{darkly_light}: Set default light theme

\item \code{darkly_dark}: Set default dark theme
}}

